- Added Mono.Nat library for UPnP and NAT-PMP automatic port forwarding. Tested working on OS X with a newer ASUS router using UPnP - Added 32bit version of ImageMagick for OS X (still not working though but at least the dylib loads now) - Started creating the Genre class for genre scanning - Fixed build error in User class

+++ WaveBox/src/DataModel/Model/User.cs
@@ -293,7 +293,7 @@ namespace WaveBox.DataModel.Model
 			}
 
 			// Compare ASCII value of each character, bitwise OR any diff
-			short status = 0;
+			int status = 0;
 			for (int i = 0; i < hash.Length; i++)
 			{
 				status |= ((int)hash[i] ^ (int)PasswordHash[i]);
